+++ doc_scraper/extensions.py
@@ -1,0 +1,5 @@
+from flask_socketio import SocketIO
+from flask_bootstrap import Bootstrap
+
+socketio = SocketIO()
+bootstrap = Bootstrap()
