+++ doc_scraper/settings.py
@@ -1,0 +1,12 @@
+from environs import Env
+
+env = Env()
+env.read_env()
+
+ENV = env.str('FLASK_ENV', default='production')
+DEBUG = ENV == 'development'
+SECRET_KEY = env.str('SECRET_KEY', default='secret!')
+DEBUG_TB_ENABLED = DEBUG
+DEBUG_TB_INTERCEPT_REDIRECTS = False
+INFO_NAMESPACE = '/info'
+IMAGE_DIR = env.str('IMAGE_DIR', default='images')
