+++ doc_scraper/views/__init__.py
@@ -1,0 +1,2 @@
+import doc_scraper.views.events
+import doc_scraper.views.main
