+++ requirements.txt
@@ -1,5 +1,6 @@
 click==6.7
 Django==1.11.13
+environs==4.0.0
 Flask==1.0.2
 flask_bootstrap
 flask_socketio==3.0.2
